Cleanup config loader, more cleanups
- Add config module loader
- Fix name of controller_registry
- Remove dependency on render function on error middleware
- Add dynamic dialect extractor to dump engine
- Move extended SQLA base types to ext module
- Move ZZZek's magic ORM code to a magic.py module
