+++ pybald/db/models.py
@@ -117,17 +117,13 @@ from sqlalchemy.ext.hybrid import (
 
 from sqlalchemy import func
 
-import json
-import pickle
-import zlib
 from pybald.db import engine, dump_engine
 from pybald.util import camel_to_underscore, pluralize
 
 from pybald import app
 # from project import mc
 
-# from sqlalchemy.orm.interfaces import SessionExtension
-from sqlalchemy.ext.mutable import Mutable
+from .ext import (ASCII, JSONEncodedDict, ZipPickler, MutationDict)
 
 from sqlalchemy import __version__ as sa_ver
 
@@ -156,235 +152,6 @@ import sqlalchemy.ext.declarative
 from sqlalchemy.ext.declarative import declarative_base
 Base = declarative_base(bind=engine)
 
-
-# # lifted from zzzeek's post on "Magic ORM"
-# # this only works for SA 7, due to the
-# # use of the event system.
-# # Source: http://techspot.zzzeek.org/2011/05/17/magic-a-new-orm/
-# # Copyright (I assume) Michael Bayer
-# # =====================================
-# from sqlalchemy.orm import (
-#             class_mapper, mapper, relationship,
-#             scoped_session, sessionmaker, configure_mappers
-#         )
-# from sqlalchemy.ext.declarative import declared_attr, declarative_base
-# from sqlalchemy import event
-# import re
-#
-# @event.listens_for(mapper, "mapper_configured")
-# def _setup_deferred_properties(mapper, class_):
-#     """Listen for finished mappers and apply DeferredProp
-#     configurations."""
-#
-#     for key, value in class_.__dict__.items():
-#         if isinstance(value, DeferredProp):
-#             value._config(class_, key)
-#
-# class DeferredProp(object):
-#     """A class attribute that generates a mapped attribute
-#     after mappers are configured."""
-#
-#     def _setup_reverse(self, key, rel, target_cls):
-#         """Setup bidirectional behavior between two relationships."""
-#
-#         reverse = self.kw.get('reverse')
-#         if reverse:
-#             reverse_attr = getattr(target_cls, reverse)
-#             if not isinstance(reverse_attr, DeferredProp):
-#                 reverse_attr.property._add_reverse_property(key)
-#                 rel._add_reverse_property(reverse)
-#
-# class FKRelationship(DeferredProp):
-#     """Generates a one to many or many to one relationship."""
-#
-#     def __init__(self, target, fk_col, **kw):
-#         self.target = target
-#         self.fk_col = fk_col
-#         self.kw = kw
-#
-#     def _config(self, cls, key):
-#         """Create a Column with ForeignKey as well as a relationship()."""
-#
-#         target_cls = cls._decl_class_registry[self.target]
-#
-#         pk_target, fk_target = self._get_pk_fk(cls, target_cls)
-#         pk_table = pk_target.__table__
-#         pk_col = list(pk_table.primary_key)[0]
-#
-#         if hasattr(fk_target, self.fk_col):
-#             fk_col = getattr(fk_target, self.fk_col)
-#         else:
-#             fk_col = Column(self.fk_col, pk_col.type, ForeignKey(pk_col))
-#             setattr(fk_target, self.fk_col, fk_col)
-#
-#         rel = relationship(target_cls,
-#                 primaryjoin=fk_col==pk_col,
-#                 collection_class=self.kw.get('collection_class', set)
-#             )
-#         setattr(cls, key, rel)
-#         self._setup_reverse(key, rel, target_cls)
-#
-# class one_to_many(FKRelationship):
-#     """Generates a one to many relationship."""
-#
-#     def _get_pk_fk(self, cls, target_cls):
-#         return cls, target_cls
-#
-# class many_to_one(FKRelationship):
-#     """Generates a many to one relationship."""
-#
-#     def _get_pk_fk(self, cls, target_cls):
-#         return target_cls, cls
-#
-# class many_to_many(DeferredProp):
-#     """Generates a many to many relationship."""
-#
-#     def __init__(self, target, tablename, local, remote, **kw):
-#         self.target = target
-#         self.tablename = tablename
-#         self.local = local
-#         self.remote = remote
-#         self.kw = kw
-#
-#     def _config(self, cls, key):
-#         """Create an association table between parent/target
-#         as well as a relationship()."""
-#
-#         target_cls = cls._decl_class_registry[self.target]
-#         local_pk = list(cls.__table__.primary_key)[0]
-#         target_pk = list(target_cls.__table__.primary_key)[0]
-#
-#         t = Table(
-#                 self.tablename,
-#                 cls.metadata,
-#                 Column(self.local, ForeignKey(local_pk), primary_key=True),
-#                 Column(self.remote, ForeignKey(target_pk), primary_key=True),
-#                 keep_existing=True
-#             )
-#         rel = relationship(target_cls,
-#                 secondary=t,
-#                 collection_class=self.kw.get('collection_class', set)
-#             )
-#         setattr(cls, key, rel)
-#         self._setup_reverse(key, rel, target_cls)
-#
-# # =====================================
-# # end of zzzeek's code for "Magic" ORM
-# # =====================================
-class ASCII(TypeDecorator):
-    '''
-    A database string type that only allows ASCII characters.
-
-    This is a data type check since all strings in the database could
-    be unicode.
-    '''
-    impl = VARCHAR
-
-    def process_bind_param(self, value, dialect):
-        '''
-        Run encode on a unicode string to make sure the string only
-        contains ASCII characterset characters. To avoid another conversion
-        pass, the original unicode value is passed to the underlying db.
-        '''
-        # run an encode on the value with ascii to see if it contains
-        # non ascii. Ignore the return value because we only want to throw
-        # an exception if the encode fails. The data can stay unicode for
-        # insertion into the db.
-        if value is not None:
-            value.encode('ascii')
-        return value
-
-    def process_result_value(self, value, dialect):
-        '''
-        Run encode on a unicode string coming out of the database to turn the
-        unicode string back into an encoded python bytestring.
-        '''
-        if isinstance(value, unicode):
-            value = value.encode('ascii')
-        return value
-
-
-class JSONEncodedDict(TypeDecorator):
-    '''
-    Represents a dictionary as a json-encoded string in the database.
-    '''
-    impl = VARCHAR
-
-    def process_bind_param(self, value, dialect):
-        '''
-        Turn a dictionary into a JSON encoded string on the way into
-        the database.
-        '''
-        if value is not None:
-            value = json.dumps(value)
-        return value
-
-    def process_result_value(self, value, dialect):
-        '''
-        Turn a JSON encoded string into a dictionary on the way out
-        of the database.
-        '''
-        if value is not None:
-            value = json.loads(value)
-        return value
-
-
-class ZipPickler(object):
-    '''Simple wrapper for pickle that auto compresses/decompresses values'''
-    def loads(self, string):
-        return pickle.loads(zlib.decompress(string))
-
-    def dumps(self, obj, protocol):
-        return zlib.compress(pickle.dumps(obj, protocol))
-
-
-class MutationDict(Mutable, dict):
-    '''
-    A dictionary that automatically emits change events for SQA
-    change tracking.
-
-    Lifted almost verbatim from the SQA docs.
-    '''
-    @classmethod
-    def coerce(cls, key, value):
-        "Convert plain dictionaries to MutationDict."
-        if not isinstance(value, MutationDict):
-            if isinstance(value, dict):
-                return MutationDict(value)
-            # this call will raise ValueError
-            return Mutable.coerce(key, value)
-        else:
-            return value
-
-    def update(self, *args, **kwargs):
-        '''
-        Updates the current dictionary with kargs or a passed in dict.
-        Calls the internal setitem for the update method to maintain
-        mutation tracking.
-        '''
-        for k, v in dict(*args, **kwargs).iteritems():
-            self[k] = v
-
-    def __setitem__(self, key, value):
-        "Detect dictionary set events and emit change events."
-        dict.__setitem__(self, key, value)
-        self.changed()
-
-    def __delitem__(self, key):
-        "Detect dictionary del events and emit change events."
-        dict.__delitem__(self, key)
-        self.changed()
-
-    def __getstate__(self):
-        '''Get state returns a plain dictionary for pickling purposes.'''
-        return dict(self)
-
-    def __setstate__(self, state):
-        '''
-        Set state assumes a plain dictionary and then re-constitutes a
-        Mutable dict.
-        '''
-        self.update(state)
 
 
 class NotFound(NoResultFound):
